LayoutTests/performance-api/performance-observer-entry-sort-expected.txt

Test PerformanceObserver mutating itself while in its callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS PerformanceObserver callback fired
PASS list.getEntries().length is 8
PASS sorted is true
PASS list.getEntries()[0].entryType is "measure"
PASS list.getEntries()[1].entryType is "measure"
PASS list.getEntries()[2].entryType is "measure"
PASS list.getEntries()[3].entryType is "mark"
PASS list.getEntries()[3].name is "mark1"
PASS list.getEntries()[7].entryType is "mark"
PASS list.getEntries()[7].name is "mark3"
PASS list.getEntries()[0].name is "measure1"
PASS list.getEntries()[1].name is "measure2"
PASS list.getEntries()[2].name is "measure3"
PASS list.getEntries()[3].name is "mark1"
PASS list.getEntries()[4].name is "mark2"
PASS list.getEntries()[5].name is "measure-matching-mark2-1"
PASS list.getEntries()[6].name is "measure-matching-mark2-2"
PASS list.getEntries()[7].name is "mark3"
PASS successfullyParsed is true

TEST COMPLETE
